refactor(client): isolate presentation state
