+++ packages/client/src/presentation/index.ts
@@ -1,0 +1,7 @@
+/** @internal */
+export {
+  PresentationState,
+  type ConversationMeta,
+  type CrossConversationEntry,
+  type CrossConvMessage,
+} from "./state.js";
