--- packages/client/src/refs.ts
@@ -1,23 +1,0 @@
-import { Effect, HashMap, Option, Ref } from "effect";
-
-/**
- * Read a `Ref` synchronously outside an Effect scope. Only safe for `Ref`s
- * that never fiber-park (the stock `Ref.Ref&lt;A>` never does). Use from
- * object methods or sync code paths that hold a `Ref` set up at construction.
- * @param ref Value supplied to the operation.
- * @returns The snapshot result.
- */
-export const snapshot = <A>(ref: Ref.Ref<A>): A => Effect.runSync(Ref.get(ref));
-
-/**
- * Lookup `key` in `m`, falling back to `dflt()` if absent. Lazy default.
- * @param m Value supplied to the operation.
- * @param key Value supplied to the operation.
- * @param dflt Value supplied to the operation.
- * @returns The get or result.
- */
-export const getOr = <K, V>(
-  m: HashMap.HashMap<K, V>,
-  key: K,
-  dflt: () => V,
-): V => Option.getOrElse(HashMap.get(m, key), dflt);
